+++ src/lib/text.rs
@@ -13,11 +13,17 @@ pub struct Merged {
     pub images: Vec<ImageSource>,
 }
 
-pub fn merge_messages(msgs: Vec<Message>, user_real_roles: bool) -> Option<Merged> {
+pub fn merge_messages(
+    msgs: Vec<Message>,
+    system: String,
+    custom_prompt: String,
+    user_real_roles: bool,
+) -> Option<Merged> {
     let line_breaks = if user_real_roles { "\n\n\x08" } else { "\n\n" };
     if msgs.is_empty() {
         return None;
     }
+    let system = system.trim().to_string();
     let size = size_of_val(&msgs);
     let mut w = String::with_capacity(size);
     let mut imgs: Vec<ImageSource> = vec![];
@@ -59,9 +65,13 @@ pub fn merge_messages(msgs: Vec<Message>, user_real_roles: bool) -> Option<Merge
         let txt = grp.into_iter().map(|m| m.1).collect::<Vec<_>>().join("\n");
         (role, txt)
     });
-    let first = msgs.next()?;
     // first message does not need prefix
-    write!(w, "{}", first.1).unwrap();
+    if !system.is_empty() {
+        w += system.as_str();
+    } else {
+        let first = msgs.next()?;
+        w += first.1.as_str();
+    }
     for (role, text) in msgs {
         let prefix = match role {
             Role::User => "Human: ",
@@ -69,13 +79,12 @@ pub fn merge_messages(msgs: Vec<Message>, user_real_roles: bool) -> Option<Merge
         };
         write!(w, "{}{}{}", line_breaks, prefix, text).unwrap();
     }
-    print_out_text(first.1.as_str(), "head.txt");
-    print_out_text(w.as_str(), "tail.txt");
+    print_out_text(w.as_str(), "paste.txt");
 
     // TODO: leave prompt empty for now
     Some(Merged {
         paste: w,
-        _prompt: String::new(),
+        _prompt: custom_prompt,
         images: imgs,
     })
 }